Add verify NFT button to /auction pages #271
